Use pointer, hash in parameters rather than the whole AST
The previous approach was confounded by
https://github.com/JuliaLang/julia/issues/10287.
To work around the limitations on the kinds of objects that
can live in parameters, I now store the AST in a field of the
object and not the parameters. To make it
retrievable by the stagedfunction (which sees only types, not
values), I use a shameless trick: store a pointer to it as
one of the parameters. This is _definitely_ a hack,
but it's a good one :-).

The hash is there to try to reduce the likelihood of a function
being redefined and the same memory address being reused; it's
a proxy for a unique identifier of the AST itself.